# perf(ext/console): break on iterableLimit & better sparse array handling