@kt3k 2022-09-19
@marcosc90 Can you add test cases in `cli/tests/unit/console_test.ts` which exercises the handling of `skipTo` values?