@marcosc90 2022-09-19
> @marcosc90 Can you add test cases in `cli/tests/unit/console_test.ts` which exercises the handling of `skipTo` values?

Added a few @kt3k 

https://github.com/denoland/deno/pull/15935/commits/230eeda27595e3d9c97beda70a4436abce19b9f0