@kt3k 2022-09-19
@marcosc90 Thanks for adding test cases!